+++ src/main/java/com/besson/arknights/block/custom/lifecycle/BunkBedInCabinTop.java
@@ -1,0 +1,24 @@
+package com.besson.arknights.block.custom.lifecycle;
+
+import com.besson.arknights.block.FurnitureBedBlock;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.util.DyeColor;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+
+public class BunkBedInCabinTop extends FurnitureBedBlock {
+
+    private static final VoxelShape SHAPE = Block.createCuboidShape(0, 0, 0, 16, 13, 16);
+
+    public BunkBedInCabinTop(DyeColor color, Settings settings) {
+        super(color, settings);
+    }
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return SHAPE;
+    }
+}