+++ src/main/java/com/besson/arknights/block/custom/lifecycle/SuckBeverages.java
@@ -1,0 +1,21 @@
+package com.besson.arknights.block.custom.lifecycle;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+
+public class SuckBeverages extends FurnitureHorizontalFacingBlock {
+    public SuckBeverages(Settings settings) {
+        super(settings);
+    }
+    private static final VoxelShape SHAPE = Block.createCuboidShape(6, 0, 6, 10, 7, 10);
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return SHAPE;
+    }
+}