+++ src/main/java/com/besson/arknights/datagen/ModRecipesProvider.java
@@ -463,15 +463,6 @@ public class ModRecipesProvider extends FabricRecipeProvider {
                 .criterion("has_item", conditionsFromTag(ItemTags.PLANKS))
                 .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "built_in_file_cabinet2"));
 
-        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.BUNK_BED_IN_CABIN, 1)
-                .pattern("#B#")
-                .pattern("AAA")
-                .input('#', Items.IRON_INGOT)
-                .input('A', ItemTags.PLANKS)
-                .input('B', ItemTags.WOOL)
-                .criterion("has_item", conditionsFromItem(Items.IRON_INGOT))
-                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "bunk_bed_in_cabin"));
-
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.CABIN_FIRE_EXTINGUISHER_KIT, 1)
                 .pattern("A")
                 .pattern("B")
@@ -479,13 +470,6 @@ public class ModRecipesProvider extends FabricRecipeProvider {
                 .input('B', Items.WATER_BUCKET)
                 .criterion("has_item", conditionsFromItem(Items.WATER_BUCKET))
                 .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "cabin_fire_extinguisher_kit"));
-
-        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.CABIN_FLUSH_MOUNT_PIPE, 1)
-                .pattern("A")
-                .pattern("A")
-                .input('A', Items.IRON_INGOT)
-                .criterion("has_item", conditionsFromItem(Items.IRON_INGOT))
-                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "cabin_flush_mount_pipe"));
 
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.CONTROL_DISPLAY, 1)
                 .pattern("###")
@@ -533,15 +517,6 @@ public class ModRecipesProvider extends FabricRecipeProvider {
                 .input('B', Items.REDSTONE)
                 .criterion("has_item", conditionsFromItem(Items.REDSTONE))
                 .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "general_console"));
-
-        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.NOTE_BOARD_IN_CABIN, 1)
-                .pattern("###")
-                .pattern("A A")
-                .pattern("###")
-                .input('#', ItemTags.PLANKS)
-                .input('A', Items.PAPER)
-                .criterion("has_item", conditionsFromItem(Items.PAPER))
-                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "note_board_in_cabin"));
 
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.BAR_CEILING, 1)
                 .pattern("###")